agent-requirements.txt
anyio
tenacity
tantivy
typer
typing_extensions
langchain-community
langchain-openai
pymupdf
requests
